restructure whole project into different folders, remove repeated .cpp files into one .lib

+++ PokerGameStaticLib/src/Deck.cpp
@@ -105,5 +105,5 @@ std::vector<Card> Deck::getDeck() const
 
 unsigned int Deck::getCardsLeft() const
 {
-	return this->deck.size();
+	return static_cast<unsigned int>(this->deck.size());
 }
